+++ demo/tests/invoice_attack_pipeline.py
@@ -37,7 +37,7 @@ from policy_registry.client import PolicyRegistryClient
 from intentframe_native_kit.resource_registry.client import ResourceRegistryClient
 from intentframe_native_kit.resource_registry.models import ResourceMount
 
-from intentframe_server.client import IntentFrameClient
+from intentframe_client import IntentFrameClient
 
 from policy_loader import load_test_policy